# [PERF] Decrease compaction RAM usage and increase speed