--- a/rust/blockstore/src/arrow/block/delta/delta.rs
+++ b/rust/blockstore/src/arrow/block/delta/delta.rs
@@ -60,7 +60,7 @@ impl BlockDelta {
         self.builder.get_size::<K>()
     }
 
-    pub fn finish<K: ArrowWriteableKey, V: ArrowWriteableValue>(&self) -> RecordBatch {
+    pub fn finish<K: ArrowWriteableKey, V: ArrowWriteableValue>(self) -> RecordBatch {
         self.builder.to_record_batch::<K>()
     }
 
@@ -172,7 +172,7 @@ mod test {
         // TODO: should commit take ownership of delta?
         // Semantically, that makes sense, since a delta is unsuable after commit
 
-        let block = block_manager.commit::<&str, Vec<i32>>(&delta);
+        let block = block_manager.commit::<&str, Vec<i32>>(delta);
         let mut values_before_flush = vec![];
         for i in 0..n {
             let key = format!("key{}", i);
@@ -208,7 +208,7 @@ mod test {
             delta.add(prefix, key.as_str(), value.to_owned());
         }
         let size = delta.get_size::<&str, String>();
-        let block = block_manager.commit::<&str, String>(&delta);
+        let block = block_manager.commit::<&str, String>(delta);
         let mut values_before_flush = vec![];
         for i in 0..n {
             let key = format!("key{}", i);
@@ -237,7 +237,7 @@ mod test {
         // test fork
         let forked_block = block_manager.fork::<&str, String>(&delta_id).await;
         let new_id = forked_block.id.clone();
-        let block = block_manager.commit::<&str, String>(&forked_block);
+        let block = block_manager.commit::<&str, String>(forked_block);
         block_manager.flush(&block).await.unwrap();
         let forked_block = block_manager.get(&new_id).await.unwrap();
         for i in 0..n {
@@ -265,7 +265,8 @@ mod test {
         }
 
         let size = delta.get_size::<f32, String>();
-        let block = block_manager.commit::<f32, String>(&delta);
+        let delta_id = delta.id.clone();
+        let block = block_manager.commit::<f32, String>(delta);
         let mut values_before_flush = vec![];
         for i in 0..n {
             let key = i as f32;
@@ -273,7 +274,7 @@ mod test {
             values_before_flush.push(read);
         }
         block_manager.flush(&block).await.unwrap();
-        let block = block_manager.get(&delta.id).await.unwrap();
+        let block = block_manager.get(&delta_id).await.unwrap();
         assert_eq!(size, block.get_size());
         for i in 0..n {
             let key = i as f32;
@@ -302,9 +303,10 @@ mod test {
         }
 
         let size = delta.get_size::<&str, RoaringBitmap>();
-        let block = block_manager.commit::<&str, RoaringBitmap>(&delta);
+        let delta_id = delta.id.clone();
+        let block = block_manager.commit::<&str, RoaringBitmap>(delta);
         block_manager.flush(&block).await.unwrap();
-        let block = block_manager.get(&delta.id).await.unwrap();
+        let block = block_manager.get(&delta_id).await.unwrap();
 
         assert_eq!(size, block.get_size());
 
@@ -366,9 +368,10 @@ mod test {
         }
 
         let size = delta.get_size::<&str, &DataRecord>();
-        let block = block_manager.commit::<&str, &DataRecord>(&delta);
+        let delta_id = delta.id.clone();
+        let block = block_manager.commit::<&str, &DataRecord>(delta);
         block_manager.flush(&block).await.unwrap();
-        let block = block_manager.get(&delta.id).await.unwrap();
+        let block = block_manager.get(&delta_id).await.unwrap();
         for i in 0..3 {
             let read = block.get::<&str, DataRecord>("", ids[i]).unwrap();
             assert_eq!(read.id, ids[i]);
@@ -400,9 +403,10 @@ mod test {
         }
 
         let size = delta.get_size::<u32, String>();
-        let block = block_manager.commit::<u32, String>(&delta);
+        let delta_id = delta.id.clone();
+        let block = block_manager.commit::<u32, String>(delta);
         block_manager.flush(&block).await.unwrap();
-        let block = block_manager.get(&delta.id).await.unwrap();
+        let block = block_manager.get(&delta_id).await.unwrap();
         assert_eq!(size, block.get_size());
 
         // test save/load
@@ -427,7 +431,7 @@ mod test {
             delta.add(prefix, key, value);
         }
         let size = delta.get_size::<u32, u32>();
-        let block = block_manager.commit::<u32, u32>(&delta);
+        let block = block_manager.commit::<u32, u32>(delta);
         let mut values_before_flush = vec![];
         for i in 0..n {
             let key = i as u32;
@@ -456,7 +460,7 @@ mod test {
         // test fork
         let forked_block = block_manager.fork::<u32, u32>(&delta_id).await;
         let new_id = forked_block.id.clone();
-        let block = block_manager.commit::<u32, u32>(&forked_block);
+        let block = block_manager.commit::<u32, u32>(forked_block);
         block_manager.flush(&block).await.unwrap();
         let forked_block = block_manager.get(&new_id).await.unwrap();
         for i in 0..n {