--- a/rust/blockstore/src/lib.rs
+++ b/rust/blockstore/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod positional_posting_list_value;
 pub mod types;
 
 pub mod arrow;