… rust/blockstore/src/arrow/sparse_index.rs
… rust/blockstore/src/memory/storage.rs
… rust/blockstore/src/positional_posting_list_value.rs
… rust/blockstore/src/types.rs
… rust/index/src/fulltext/types.rs
… rust/worker/src/segment/metadata_segment.rs