## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
	 - BTree of block deltas for postings list is now Vec<i32> instead of IntArray since the overhead of IntArray was seen to be 2 KB per btree add v/s 816 bytes for Vec<i32>
	 - When getting blocks from block deltas at commit time, the block deltas are now drained. This reduces RAM consumption by about 50%
	 - Removed the intermediary postings list builder that was more of an unnecessary complexity and also was doing clones
	 - Changed get() for postings list to read slices instead of Vec<i32> getting rid of one more deep copy.

## Test plan
*How are these changes tested?*
- [x] Tests pass locally with `pytest` for python, `yarn test` for js, `cargo test` for rust

## Documentation Changes
None