@codetheweb 2024-08-29
Perf results, comparing the latest in this branch against https://github.com/chroma-core/chroma/pull/2729/commits/14a744bce2b70fcf458de1adb68c203fd739c174 with the other [tokenizer](https://github.com/chroma-core/chroma/pull/2735) and [locking updates](https://github.com/chroma-core/chroma/pull/2736):

## Time profiling

Follows the same experiment structure as https://github.com/chroma-core/chroma/pull/2736.

[sanket-branch-final-cpu.trace.zip](https://github.com/user-attachments/files/16790371/sanket-branch-final-cpu.trace.zip)

**Before** (see trace in https://github.com/chroma-core/chroma/pull/2736):

- First compaction: 33s
- Second compaction: 40s

**After** (run 1):

- First compaction: 29s
- Second compaction: 36s

## Memory profiling

https://wormhole.app/Z26BY#aW8Ij5sW14ORvbbGpkPXiw

1. Added 5,000 documents
2. Compacted
3. Added 5,000 documents
4. Compacted

**Before** (run 2):

- First compaction peak: 600 MiB
- Second compaction peak: 1146  MiB

**After** (run 1):

- First compaction peak:  523 MiB
- Second compaction peak: 1064 MiB